RemoteLinux: Handle already remote local files
This is a bit hackish to support "true" remote builds.

Conceptually, that's more similar to a "local" (a.k.a.
"Desktop" configurations when it comes to (non-)deployment,
but the Desktop runconfigs are currently a bit too tied
to the host system, and we'd like to have the X11 forwarding
that's already available here.

Change-Id: I750b2cfe6b94ec8f60a0e879f3fe44ddaf86072e
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>