qmlpuppet: Fix build with Qt 5
The API uses QVector, which is a QList in Qt 6, which is why assigning a
QVector to a QList works "fine" there. But with Qt 5 we must use the
correct type.

Also assign to a const variable. Otherwise this actually forces a
detach.

Amends 9321c6ad24042a256ac44d28e5263e45f0fb4a16

Change-Id: Ied562b96d647435ae48395ae08cc12a5670ac607
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Artem Sokolovskii <artem.sokolovskii@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>