JSON wizards: Fix "span" with combo boxes
Combo boxes have "suppressName" set, but if they have span set too, we
still need to use addRow instead of addWidget.

Reorder the if conditions. The previous code didn't make much sense,
because it first checked for suppressName, and then for hasSpan, but
inside hasSpan it checked for (!suppressName) again, which could never
be false.

You can see this in action in the File > Qt > JS File wizard.

Change-Id: I976227f80277642823620e41aba605c8570aed44
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>