Python: Remove wizard for dynamically loaded ui projects
The dialog and main window variant of this wizard was broken since ever.
Generally two widget ui templates are generally confusing for the user,
so remove the broken wizard that promotes the not recommended way of
using ui files. Also we support the automatic conversion from ui to py
files nowadays via the PySide6 project tool.

Fixes: QTCREATORBUG-25807
Change-Id: I3ec6000a6c3fb076e3a54c70f3cccdd12df28701
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>