Squish: Add common start function to process base
Starting the process is similar for the tools, so extract it
to the base class.
Make it virtual to let the tools still be able to re-init internal
members.
Slightly changes the startup of the server as the Started state
was emitted later manually after the port in use had been known.

Change-Id: Iea88ce88e8a12a7723460adaa25f7b9a5cd9a2b8
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>