QmlDesigner: Fix value types in property editor
Currently the value types of e.g. layer.textureSize and layer.sourceRect
are only resolved for QML Item, but not QML Control. This patch
increases the recursion depth while collecting the attributes of values
types in order to be able to set layer.textureSize.width or
layer.sourceRect.x.
It also adds those attributes to be able to read the values.

Change-Id: I61ba1468d1443953f0a5b6ab2241114dc441bb79
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>