QmlDesigner fix: Project style is configured wrong when filtered
In the "New Project" dialog box, if the "Style" section is filtered (to
showing e.g. only "Light" or only "Dark" themes) then the project being
created was configured with the wrong style. The problem was that the
index of the style was not "translated" from the filtered list to
normal list. This should have caused problems when saving the preset as
well, if the "Styles" section was filtered (to white/dark).

Task-number: QDS-8874
Change-Id: I9422dad3083f9bac3355bd65e1a71332d093271d
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>