QmlDesigner: Make the selected style appear selected even when hovered
In the New Project dialog, when browsing thorough the Styles, the
border of each style becomes white on hover, and when selected, the
border is blue. When a Style item was both hovered and selected,
though, the border was shown as white (i.e. highlighted). It should
be blue (selected) instead.

Task-number: QDS-8875
Change-Id: I4573ff57b7b5561f17f1fe6adf3fa214a62d5538
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>