QmlDesigner: Ensure Effects folder is visible after recreate
Moved the Effect Maker functionality from assetslibrary model to the
widget, so as to have access to the `directoryCreated()` which is used
on the QML side to (forcibly) update the rows. This should fix it, even
though I could not reproduce the problem for effects myself -- adding
ordinary assets, though, always emits `directoryCreated()` to avoid
this kind of problem.

Task-number: QDS-8916
Change-Id: Ib72c4a235851eed266b6751bb67d4de6acbcfcc5
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>