QmlDesigner: Fix renaming a folder in Assets View expands the wrong folder
The previous implementation was saving the _row_ of the TreeView before
performing the rename, and after the rename it was expanding that _row_.
However, when renaming a folder causes it to be moved up or down between
the rows, the saved _row_ number becomes invalid.

The better way is just to use the same mechanism that is being used when
creating a new folder: add the new folder path to a list, and later on
expand rows based on that list.

Task-number: QDS-8896
Change-Id: I6a6bb2faaa7d03ea2af384e5f2b635000577a94c
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>