Squish: Fix cancel handling
Remove canceling as state as it never had been used for real.
Keep the canceled state instead of overwriting it with finished.
Canceled has internally anyhow the same meaning as finished,
but we need the information whether the process had been canceled
or not to handle secondary runners like the recorder correctly.
Fixes canceling a run or a record.

Change-Id: I20ff6bc1b1d0b969c19940f763411e26149b3d16
Reviewed-by: David Schulz <david.schulz@qt.io>