QmlDesigner: Fix condition chains in 3D import options
When an import option conditionally depends on another option that
depends on a third option, toggling the third option needs to change
the enabled state of the first option controls as well as the second
option controls.

For example, consider this set of options (actual case with options
for generating mesh levels of detail in Qt 6.5):

optA
  optB, enabled when optA is true
    optC, enabled when optB is true
    optD, enabled when optB is true

If optA is false, optC and optD controls need to be disabled even if
optB is true, because of the chained conditions.

Fixes: QDS-9055
Change-Id: I10aeb7596cf9c8a9053c7d08ec9ebb7a46db2cd2
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>