ExtensionSystem: Another way to have plugin tests
This makes it possible to have plugin tests defined in the
plugin pimpl, like a lot of the other stuff we currently
have there. This is meant to replace IPlugin::createTestObjects.

The basic idea is to reduce the number of ways things can
or have to be done when creating plugins.

Change-Id: I137c66b91cdf461a502cf22063b6f63e6163e876
Reviewed-by: Eike Ziller <eike.ziller@qt.io>