McuSupport: Expand path macros when resetting paths
Path macros were expanded when applying new paths in the
McuSupport options widget, but not when pushing the reset button.
This resulted in paths like {Env:PROGRAMFILES}/SEGGER/JLink,
which are displayed as an error in the UI. The reset-button is
now connected via a signal in McuPackage to the options widget,
since the macro expander is not accessible to the package object
directly.

Task-number: QTCREATORBUG-28500
Change-Id: I962c6eb83142be87e02debbf34f0602683fe6966
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Yasser Grimes <yasser.grimes@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>