McuSupport: Discard un-applied changes to Qul SDK package
The path to the Qul SDK entered in the settings path persisted in
the UI when re-opening the settings page, even if it was not
applied. Kits and the settings file are not affected unless
pressing apply so this did not affect functionality. Un-applied
changes are discarded by re-reading the path from the settings
file in the constructor for the settings page.

Task-number: QTCREATORBUG-28568
Change-Id: Id03be22fba429e4764493cd3bedf45faa6908637
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Yasser Grimes <yasser.grimes@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>