ProjectExplorer: Do not expand the current entry in the issues pane
The expansion of the current item makes for a "jumpy" user experience
and also runs into a nasty Qt bug when navigating via keyboard.
Instead, we now show all items the same way and provide the details in a
tool tip.
As a side effect, build issues in text marks are now linkified.

Fixes: QTCREATORBUG-26128
Fixes: QTCREATORBUG-27006
Fixes: QTCREATORBUG-27506
Task-number: QTBUG-100309
Change-Id: Id9078c6f798fad31781bb645e06bbc92fc73fba5
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>