Link with Qt: Remove now harmful check
The check was trying to find out if the Qt Creator installation is part
of the online installer or a Qt installer, by checking if there are
install settings without the key "Settings/InstallSettings".

Since we made Qt Creator optional in the installer, the install settings
contain that redirection key anyway, so it is no longer an indicator for
Qt Creator being part of the online or an Qt installer anymore.

Actually the check now _disables_ the Link with Qt button for the
commercial Qt Creator offline installer on macOS, because that ships
install settings that add the plugin path to the commercial plugins
outside the Qt Creator app bundle, but does not contain "Settings/
InstallSettings".

Just remove the check.

Change-Id: I88c3ef007a728dff0f4214e5a5683ffcede23796
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: David Schulz <david.schulz@qt.io>