Core: reduce flickering when restarting a search
The cancel button of the search result widget gets visible when a search
is currently ongoing in other states it is hidden. Showing and hiding
the cancel button changes the height of the replace tool bar. In the
case of clangd a search is started whenever the replacement text changes
to update the results. So if the results are received shortly after
requesting them the cancel button is only shown briefly resulting in the
flickering. Avoid this by setting a fixed minimum height to the label
containing the cancel button.

Change-Id: I6fb7a4cfe256c7612a902be621fa9ad971681ee7
Reviewed-by: Eike Ziller <eike.ziller@qt.io>