AutoTest: Remove gtest notes from visual tree
GTest uses a linear output format while we use a tree to
group the results.
Some information which make sense for the linear result
cannot be placed correctly into the visual tree at all
without breaking either the tree layout or displaying the
result at unexpected locations inside the tree.
So, omitting these information to avoid confusion seems
to be the best trade-off as the information is still
present inside the text output if needed.

Change-Id: Ifd72bbf7e88993b4c76765597288e1531b50ee9a
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: David Schulz <david.schulz@qt.io>