Fix vanishing search term from locator popup
When setting the locator to open as a centered popup, everything locator
is triggered, the search term is cleared.

Don't delete the popup when closing it. That keeps the input field alive,
and the text in it. We manage a singleton instance of it in
locatormanager.cpp/locatorWidget(), and it will eventually be delete
because it has a parent.

Change-Id: I237029bc2793f7394c5eb6ea7d658a4dd4a2c1cf
Reviewed-by: Artem Sokolovskii <artem.sokolovskii@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: David Schulz <david.schulz@qt.io>