QmlDesigner: Force QDS compatibility mode 3D import option
QDS requires 3D imports to use QDS compatibility mode to work correctly,
so we force designStudioWorkarounds option on in all imports we do and
hide the option from UI so user can't toggle it off.
Since designStudioWorkarounds supercedes expandValueComponents option,
we also hide that as setting it has no effect.

Fixes: QDS-9219
Change-Id: Id40a45338ea630228f59b111703eba661ebd357b
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>