Debugger: Clear also global log when clearing the log
If the user wants to clear the log in order to read it only "from now
on", then the sensible way to do that is to clear the log, finish the
debugging session, then copy the output from the log pane.

But this flow did not work, because when the debugger is done, you no
longer have access to the live log window, only to the global window,
and this pane is not cleared, so all the logs are still there.

Change-Id: I89d8a5b8aab22c53e8c2eceac4c5482931cdd8b6
Reviewed-by: hjk <hjk@qt.io>