GDB: Do not skip detach for core inferior shutdown
Another attempt for 87c00b86dbf75a9204041b0b8e836415b97dc125.

This section was added in 3b5ecac238, which was a refactoring commit,
not sure why.

It causes state discrepancy, because only on quit the "thread-exited"
messages are processed, and when the thread list is clear, the engine
considers it as InferiorShutdownFinished and triggers engine shutdown
again, which is wrong, because it quits twice and can no longer write
to the process, which is already dead at this point.

Change-Id: Ic5cfced672b640865e701ac96b6fcea34d748211
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>