FileIconProvider: Improve performance of icon(FilePath)
By re-using the caching QFileInfo variant for icon(FilePath).
QFileInfo/FSEngine caches the result of isDir etc for a while, which
improves performance for slower devices in case that is called for the
same file paths over and over again, like it is the case for locator.

The cache might be out of date for some time when things change on disk,
but for a cosmetic property like an icon that is not a big deal. (And it
would be a file that transforms into a directory or vice versa, which is
not very common either.)

Change-Id: I663ac7e81f8f2996b87591dad17c5483e3f4ad3c
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>